LayoutTests/discovery/discovery-zeroconf-api-expected.txt

Add request
WebView network services request permission 0 service(s) (origin "file://")
	Services:
	 - 0 allowed
	 - 0 denied
	 - 0 online
	 - 0 offline
Tests that getNetworkServices API is available and support all characters for valid service type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.getNetworkServices is defined.
* register default error callback
* sent network services request
PASS successfullyParsed is true

TEST COMPLETE
* got reply: 0 service(s) on 0 services available
